% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_functions.R
\name{calculate_limits}
\alias{calculate_limits}
\title{Calculate the space restriction limits of a parameter}
\usage{
calculate_limits(comps, vTs)
}
\arguments{
\item{comps}{result of \code{calculate_components} function}

\item{vTs}{named list of testvectors}
}
\value{
named \code{list} of class \code{limitObject}. Each element corresponds to one 
test vector and does contain \code{vT}, the test vector, and \code{limits}, 
an \code{Intervals} object for the calculated limits
}
\description{
Calculate the space restriction limits of a parameter
}
