% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-value_calculation.R
\name{ci_tnorm}
\alias{ci_tnorm}
\title{Calculate confidence interval for truncated normal variable}
\usage{
ci_tnorm(meanEst, sd, limits, alpha, gridpts = 500, griddepth = 3,
  range = 100)
}
\arguments{
\item{meanEst}{mean}

\item{sd}{standard deviation}

\item{limits}{\code{Intervals} object containing all limits}

\item{alpha}{value specifying the interval coverage, which is \code{(1-alpha)}}

\item{gridpts, griddepth}{parameters for grid search used in the confidence interval
function of the package \code{selectiveInference}.}

\item{range}{gives the range \code{range*sd}, in which the quantile values are searched}
}
\description{
Calculate confidence interval for truncated normal variable
}
\examples{
library(intervals)
# one restriction -- in this case equal to a call to \\code{\\link{msm::qtnorm}}
int1 <- Intervals(c(-3,1))
mean = -0.2
sd = 1
ci_tnorm(meanEst = mean, sd = sd, limits = int1, alpha = 0.05)

# two intervals -- 1-alpha coverage not possible
(int2 <- Intervals(matrix(c(-3,-2,-1,1), 
                          ncol = 2, byrow = TRUE)))
ci_tnorm(meanEst = mean, sd = sd, limits = int2, alpha = 0.05)

# two intervals -- 1-alpha coverage possible, but interval not symmetric
(int3 <- Intervals(matrix(c(-3,-2.5,-1.9,1), 
                          ncol = 2, byrow =  TRUE)))
ci_tnorm(meanEst = mean, sd = sd, limits = int3, alpha = 0.05)

}
