% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_functions.R
\name{extract_testvec}
\alias{extract_testvec}
\title{extract all test vectors from \code{lm} object}
\usage{
extract_testvec(limo)
}
\arguments{
\item{limo}{\code{lm} object}
}
\value{
\code{list} of test vectors
}
\description{
extract all test vectors from \code{lm} object
}
\examples{
set.seed(42)
y <- rnorm(10)
x <- runif(10)
mod <- lm(y ~ x)

# get test vectors
vT <- extract_testvec(mod)
# calculate coefficient estimates
sapply(vT, function(vt) vt\%*\%y)
coef(mod)

}
