% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_functions.R
\name{getBoundsPen}
\alias{getBoundsPen}
\title{Function to calculation boundaries for given components}
\usage{
getBoundsPen(A, pv, y, vt, c = 0)
}
\arguments{
\item{A}{restriction matrix in affine inequality}

\item{pv}{projection matrix for test vector or its left singular values for group variables}

\item{y}{response}

\item{vt}{test vector}

\item{c}{additive value in affine inequality; defaults to 0}
}
\description{
Function to calculation boundaries for given components
}
