% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_functions.R
\name{getHatMat}
\alias{getHatMat}
\title{hat matrix calculation of linear model}
\usage{
getHatMat(limo = NULL, X = NULL)
}
\arguments{
\item{limo}{\code{lm} object}

\item{X}{\code{matrix} respresenting the design matrix of a linear model}
}
\value{
hat matrix
}
\description{
hat matrix calculation of linear model
}
\examples{
set.seed(42)
y <- rnorm(10)
x <- runif(10)
mod <- lm(y ~ x)

str(getHatMat(mod),1)
str(getHatMat(X = as.matrix(cbind(interc=rep(1,10),x))),1)

}
