% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selinf.R
\name{selinf}
\alias{selinf}
\title{Calculate p-values / confidence intervals after likelihood-based or test-based model selection}
\usage{
selinf(listOfModels, ..., response = NULL, what, REML = FALSE, df = 1, sd,
  alpha = 0.05, gridpts = 500, griddepth = 3, range = 100)
}
\arguments{
\item{listOfModels}{a \code{list} of models representing one comparison}

\item{...}{optionally more \code{list}s of models representing further comparisons}

\item{response}{response vector}

\item{what}{\code{character} describing the method, on the basis of which variable selection was performed.
Supported variable selection procedures are AIC (\code{"aic"}), BIC (\code{"bic"}), 
likelihood-based selection (\code{"llonly"}) and significance tests based on the likelihood-ratio 
or the F-statistic (\code{"lrt"}, \code{"Ftest"}). If different selection procedures have been used, 
please supply the corresponding character for each list of models.}

\item{REML}{\code{logical}; whether scale estimates should be calculated via \code{REML} or not. Default is
\code{FALSE} since the log-likelihood is calculated with \code{REML=FALSE} by the \code{logLik} function.}

\item{df}{positive integer; defines the degree of freedom when using the likelihood ratio test for model selection.
Defaults to \code{1} (testing one parameter at a time).}

\item{sd}{standard deviation of error used for the p-value calculation (see details)}

\item{alpha}{value for \code{(1-alpha)}-confidence interval construction. Defaults to \code{0.05}.}

\item{gridpts, griddepth, range}{options for the calculation of confidence intervals. 
See \code{\link{ci_tnorm}} for more details.}
}
\description{
This is a wrapper and convenience function for several functions included in this package. 
Please see \code{\link{calculate_selinf}} for more details and how to calculate inference when
several model selection procedures are combined. 
When more than one list is supplied, the best model in the first \code{listOfModels} is used to perform inference.
}
\examples{
library(MASS)
data("cpus")
# Fit initial model
cpus$perf <- log10(cpus$perf)
cpus$cach <- as.factor(cpus$cach)
mod <- lm(perf ~ .-name, data = cpus)

# use the stepAIC function to find the best model in a backward 
# stepwise search
cpus.lm <- stepAIC(mod, trace = FALSE, direction = "backward", steps = 3)
# check model selection
cpus.lm$anova$Step

# recalculate all visited models in the first step
lom1 <- c(lapply(attr(mod$terms, "term.labels"), function(x) 
update(mod, as.formula(paste0("perf ~ .-", x)))), list(mod))

# perform likelihood ratio test at level 
alpha = 0.001

# check for non-significant variables
coefTable <- anova(cpus.lm)
drop <- rownames(coefTable)[alpha < coefTable[-nrow(coefTable),5]]

# drop non-significant variable
cpus.lm2 <- update(cpus.lm, as.formula(paste0(".~.-",drop)))

# compute selective inference
selinf(list(cpus.lm, cpus.lm2), lom1, 
response = cpus$perf,
what = c("Ftest", "aic"),
sd = summary(cpus.lm2)$sigma)

}
