% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_functions.R
\name{extract_components}
\alias{extract_components}
\title{extract components from list of \code{lm} objects}
\usage{
extract_components(listOfModels, response = NULL, what = c("aic", "bic",
  "llonly", "lrt", "Ftest", "sigHunt"), REML = FALSE, alpha = NULL,
  df = 1)
}
\arguments{
\item{listOfModels}{\code{list} of \code{lm} objects}

\item{response}{response vector; must be supplied if not stored in the \code{lm} object.}

\item{what}{\code{character} describing the method, on the basis of which variable selection was performed}

\item{REML}{\code{logical}; whether scale estimates should be calculated via \code{REML} or not. Default is
\code{FALSE} since the log-likelihood is calculated with \code{REML=FALSE} by the \code{logLik} function.}

\item{alpha}{significance level for selection via likelihood ratio test or significance hunting.}

\item{df}{positive integer; defines the degree of freedom when using the likelihood ratio test for model selection.
Defaults to \code{1} (testing one parameter at a time).}
}
\value{
\code{list} of components for each model. \code{A} and \code{c}: matrices and scalar of 
affine inequality; \code{y}: response vector; \code{bestInd} indicator for best model in \code{listOfModels}.
}
\description{
extract components from list of \code{lm} objects
}
\examples{
set.seed(42)
y <- rnorm(10)
x <- runif(10)
mod1 <- lm(y ~ x)

x2 <- runif(10)
mod2 <- lm(y ~ x2)

# AIC comparison
AIC(mod1,mod2)

lom <- list(mod1, mod2)

# get components
comps <- extract_components(lom, response = y)
str(comps,1)
t(comps$y)\%*\%comps$A[[1]]\%*\%comps$y


}
