% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_functions.R
\name{getBoundsPen}
\alias{getBoundsPen}
\title{Function to calculation boundaries for given components}
\usage{
getBoundsPen(p1 = NULL, p2 = NULL, Px1, Px2, pen1, pen2, pv, y, vt, REML)
}
\arguments{
\item{p1, p2}{number of parameters in model 1 and 2}

\item{Px1, Px2}{hat matrix of model 1 and 2}

\item{pen1, pen2}{penalty of model 1 and 2}

\item{pv}{projection matrix for test vector}

\item{y}{response}

\item{vt}{test vector}

\item{REML}{logical; whether model variance is estimated using \code{REML}. 
If \code{REML = TRUE}, \code{p1} and \code{p2} must be supplied.}
}
\description{
Function to calculation boundaries for given components
}
