% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selinf.R
\name{selinf}
\alias{selinf}
\title{Calculate p-values / confidence intervals after likelihood-based or test-based model selection}
\usage{
selinf(listOfModels, ..., response = NULL, what = c("aic", "bic", "llonly",
  "lrt", "Ftest"), REML = FALSE, df = 1, sd, alpha = 0.05)
}
\arguments{
\item{listOfModels}{a \code{list} of models representing one comparison}

\item{...}{optionally more \code{list}s of models representing further comparisons}

\item{response}{response vector}

\item{what}{\code{character} describing the method, on the basis of which variable selection was performed.
If different selection procedures have been used, please supply the corresponding character for each
list of models.}

\item{REML}{\code{logical}; whether scale estimates should be calculated via \code{REML} or not. Default is
\code{FALSE} since the log-likelihood is calculated with \code{REML=FALSE} by the \code{logLik} function.}

\item{df}{positive integer; defines the degree of freedom when using the likelihood ratio test for model selection.
Defaults to \code{1} (testing one parameter at a time).}

\item{sd}{standard deviation of error used for the p-value calculation (see details)}

\item{alpha}{value for \code{(1-alpha)}-confidence interval construction. Defaults to \code{0.05}.}
}
\description{
This is a wrapper and convenience function for several functions included in this package. 
Please see \code{\link{calculate_selinf}} for more details and how to calculate inference when
several model selection procedures are combined. 
When more than one list is supplied, the best model in the first \code{listOfModels} is used to perform inference.
}
